+++ apps/meteor/app/importer/server/classes/converters/MessageConverter.ts
@@ -38,7 +38,7 @@ export class MessageConverter extends RecordConverter<IImportMessageRecord> {
 	}
 
 	protected async resetLastMessages(): Promise<void> {
-		for await (const rid of this.rids) {
+		for (const rid of this.rids) {
 			try {
 				await Rooms.resetLastMessageById(rid, null);
 			} catch (err) {
@@ -120,7 +120,7 @@ export class MessageConverter extends RecordConverter<IImportMessageRecord> {
 		}
 
 		const result: MentionedChannel[] = [];
-		for await (const importId of channels) {
+		for (const importId of channels) {
 			const { name, _id } = (await this.getMentionedChannelData(importId)) || {};
 
 			if (!_id || !name) {
@@ -146,7 +146,7 @@ export class MessageConverter extends RecordConverter<IImportMessageRecord> {
 		}
 
 		const result: MentionedUser[] = [];
-		for await (const importId of mentions) {
+		for (const importId of mentions) {
 			if (importId === ('all' as 'string') || importId === 'here') {
 				result.push({
 					_id: importId,
@@ -185,7 +185,7 @@ export class MessageConverter extends RecordConverter<IImportMessageRecord> {
 	): Promise<undefined | IMessageReactions> {
 		const reactions: IMessageReactions = {};
 
-		for await (const name of Object.keys(importedReactions)) {
+		for (const name of Object.keys(importedReactions)) {
 			if (!importedReactions.hasOwnProperty(name)) {
 				continue;
 			}
@@ -200,7 +200,7 @@ export class MessageConverter extends RecordConverter<IImportMessageRecord> {
 				usernames: [],
 			};
 
-			for await (const importId of users) {
+			for (const importId of users) {
 				const username = await this._cache.findImportedUsername(importId);
 				if (username && !reaction.usernames.includes(username)) {
 					reaction.usernames.push(username);
@@ -219,7 +219,7 @@ export class MessageConverter extends RecordConverter<IImportMessageRecord> {
 
 	protected async convertMessageReplies(replies: string[]): Promise<string[]> {
 		const result: string[] = [];
-		for await (const importId of replies) {
+		for (const importId of replies) {
 			const userId = await this._cache.findImportedUserId(importId);
 			if (userId && !result.includes(userId)) {
 				result.push(userId);