+++ apps/meteor/packages/rocketchat-mongo-config/server/index.js
@@ -4,13 +4,8 @@ import { PassThrough } from 'stream';
 import { Email } from 'meteor/email';
 import { Mongo } from 'meteor/mongo';
 
-const shouldUseNativeOplog = ['yes', 'true'].includes(
-	String(process.env.USE_NATIVE_OPLOG).toLowerCase(),
-);
-
-if (!shouldUseNativeOplog) {
-	Package['disable-oplog'] = {};
-}
+// we always want Meteor to disable oplog tailing
+Package['disable-oplog'] = {};
 
 // FIX for TLS error
 // https://github.com/RocketChat/Rocket.Chat/issues/9316