+++ apps/meteor/server/services/device-management/events.ts
@@ -1,6 +1,5 @@
-import type { ISocketConnectionLogged } from '@rocket.chat/core-typings';
 import { Emitter } from '@rocket.chat/emitter';
 
 export const deviceManagementEvents = new Emitter<{
-	'device-login': { userId: string; connection: ISocketConnectionLogged };
+	'device-login': { userId: string; userAgent: string; clientAddress: string };
 }>();