+++ packages/core-typings/src/mediaCalls/IMediaCall.ts
@@ -65,4 +65,7 @@ export interface IMediaCall extends IRocketChatRecord {
 	transferredAt?: Date;
 
 	uids: IUser['_id'][];
+
+	/** The list of features that may be used in this call. Values are final once the call is accepted. */
+	features: string[];
 }